compute_rf(::PEP...
